Move createPoll logic to MessageComposerController
