+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/ui/MessagesActivity.kt
@@ -251,7 +251,7 @@ class MessagesActivity : BaseConnectedActivity() {
                     onTabClick = { _, tab -> isFullScreenContent = tab.isFullContent },
                     onAttachmentPickerAction = { action ->
                         if (action is AttachmentPickerPollCreation) {
-                            listViewModel.createPoll(
+                            composerViewModel.createPoll(
                                 pollConfig = PollConfig(
                                     name = action.question,
                                     options = action.options.filter { it.title.isNotEmpty() }.map { it.title },
