+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/messages/MessageComposerViewModel.kt
@@ -22,6 +22,7 @@ import io.getstream.chat.android.models.ChannelCapabilities
 import io.getstream.chat.android.models.Command
 import io.getstream.chat.android.models.LinkPreview
 import io.getstream.chat.android.models.Message
+import io.getstream.chat.android.models.PollConfig
 import io.getstream.chat.android.models.User
 import io.getstream.chat.android.ui.common.feature.messages.composer.MessageComposerController
 import io.getstream.chat.android.ui.common.state.messages.Edit
@@ -170,6 +171,15 @@ public class MessageComposerViewModel(
      */
     public fun removeSelectedAttachment(attachment: Attachment): Unit =
         messageComposerController.removeSelectedAttachment(attachment)
+
+    /**
+     * Creates a poll with the given [pollConfig].
+     *
+     * @param pollConfig Configuration for creating a poll.
+     */
+    public fun createPoll(pollConfig: PollConfig) {
+        messageComposerController.createPoll(pollConfig = pollConfig)
+    }
 
     /**
      * Sends a given message using our Stream API. Based on the internal state, we either edit an existing message,
