+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/messages/MessageListViewModel.kt
@@ -25,7 +25,6 @@ import io.getstream.chat.android.models.ConnectionState
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.Option
 import io.getstream.chat.android.models.Poll
-import io.getstream.chat.android.models.PollConfig
 import io.getstream.chat.android.models.User
 import io.getstream.chat.android.models.Vote
 import io.getstream.chat.android.state.plugin.state.channel.thread.ThreadState
@@ -431,15 +430,6 @@ public class MessageListViewModel(
      */
     public fun performGiphyAction(action: GiphyAction) {
         messageListController.performGiphyAction(action)
-    }
-
-    /**
-     * Creates a poll with the given [pollConfig].
-     *
-     * @param pollConfig Configuration for creating a poll.
-     */
-    public fun createPoll(pollConfig: PollConfig) {
-        messageListController.createPoll(pollConfig = pollConfig)
     }
 
     /**
